my-lambda-layer/requirements.txt
numpy==1.22.4
pandas==1.4.0
requests==2.28.0
boto3==1.9.4
toml==0.10.2
python-dotenv==0.20.0
SQLAlchemy==1.4.39
mysql-connector-python==8.0.29